+++ cad/src/lib.rs
@@ -17,7 +17,7 @@ pub struct Plane {
 #[derive(Debug, Serialize, Deserialize)]
 pub struct SketchInfo {
     /// Stable id that doesnt change even if ordering does in a scene
-    pub id: i32,
+    pub id: u16,
     pub plane: Plane,
     pub sketch: Sketch,
     pub name: String,
